requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
fastapi==0.97.0
uvicorn==0.22.0
joblib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
pydantic==1.10.9
nltk==3.8.1
